Changed Rendering of custom buttons.
Rendering of custum buttons in a form has been removed from the FormRenderer.
The FormRenderer now only renders the default buttons.
Instead of this custom form rendering is now also done in form.mako and is now
part of the normal rendering flow. As now the rendering is done in form.mako
conditionals also work for buttons.